# api-contracts/full-api-contract.yaml
openapi: 3.1.0
info:
  title: Agentic Analytics Platform - Full API Contract
  description: |
    Complete API contract for the Agentic Analytics Platform.
    Covers Phase 1 (Foundation) through Phase 8 (API Integration).

    ## Authentication
    - JWT Bearer token authentication
    - OAuth2/OIDC support (Azure AD, Google)
    - API Key authentication for programmatic access
    - Token refresh mechanism

    ## Authorization
    - Role-Based Access Control (RBAC)
    - Permission-based resource access
    - Predefined roles: SuperAdmin, Admin, Analyst, Viewer, DataManager

    ## Real-time Features
    - WebSocket endpoints for voice, queries, notifications, and live dashboards

    ## Rate Limiting
    - Per-user and per-API-key rate limiting
    - Configurable limits per endpoint group
  version: 2.0.0
  contact:
    name: API Support
    email: support@agentic-analytics.com
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://api.agentic-analytics.com/api/v1
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register

    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/common/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/common/oauth2/v2.0/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: User email

    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for programmatic access
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Error message
          example: "Invalid credentials"
        error_code:
          type: string
          description: Machine-readable error code
          example: "AUTH_INVALID_CREDENTIALS"
        metadata:
          type: object
          nullable: true
          description: Additional error context
    ValidationErrorResponse:
      type: object
      properties:
        detail:
          type: string
          example: "Validation failed"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
              code:
                type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"
...
    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email:
          type: string
          format: email
          example: "john.doe@company.com"
        password:
          type: string
          minLength: 8
          example: "SecurePass123!"
        full_name:
          type: string
          minLength: 2
          example: "John Doe"
        phone:
          type: string
          example: "+1-555-123-4567"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        remember_me:
          type: boolean
          default: false
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserResponse'
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
        new_password:
          type: string
          minLength: 8
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, new_password]
      properties:
        token:
          type: string
        new_password:
          type: string
          minLength: 8
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
          format: email
        full_name:
          type: string
        phone:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        auth_provider:
          type: string
          enum: [local, azure_ad, google]
        is_active:
          type: boolean
        mfa_enabled:
          type: boolean
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponse'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        is_system:
          type: boolean
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/PermissionResponse'
        user_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    PermissionResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        resource:
          type: string
        action:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    QueryRequest:
      type: object
      required: [database_id, natural_language]
      properties:
        database_id:
          type: integer
        natural_language:
          type: string
          example: "Show me the top 10 customers by revenue in the last quarter"
        conversation_id:
          type: integer
          nullable: true
        options:
          type: object
          properties:
            max_rows:
              type: integer
              minimum: 1
              maximum: 10000
              default: 100
            timeout_seconds:
              type: integer
              minimum: 5
              maximum: 300
              default: 30
            include_metadata:
              type: boolean
              default: false
    QueryResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, executing, completed, failed, cancelled]
        natural_language:
          type: string
        generated_sql:
          type: string
          nullable: true
        explanation:
          type: string
          nullable: true
        results:
          type: object
          nullable: true
          properties:
            columns:
              type: array
              items:
                type: string
            rows:
              type: array
              items:
                type: array
            row_count:
              type: integer
            execution_time_ms:
              type: integer
        suggested_visualizations:
          type: array
          items:
            $ref: '#/components/schemas/VisualizationSuggestion'
        tokens_used:
          type: integer
        conversation_id:
          type: integer
        parent_query_id:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
    QueryListResponse:
      type: object
      properties:
        queries:
          type: array
          items:
            $ref: '#/components/schemas/QueryResponse'
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        pages:
          type: integer
    SQLExecutionRequest:
      type: object
      required: [database_id, sql]
      properties:
        database_id:
          type: integer
        sql:
          type: string
          example: "SELECT COUNT(*) FROM customers WHERE created_at > '2026-01-01'"
        timeout_seconds:
          type: integer
          default: 30
    FollowUpRequest:
      type: object
      required: [natural_language]
      properties:
        natural_language:
          type: string
          example: "What about last month?"
    VisualizationSuggestion:
      type: object
      properties:
        type:
          type: string
          enum: [bar_chart, line_chart, pie_chart, table, kpi, scatter_plot, area_chart, heatmap, histogram]
        title:
          type: string
        config:
          type: object

    # ============================================
    # PHASE 3: CONVERSATION SCHEMAS
    # ============================================
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
...
    QueryIdPath:
      name: query_id
      in: path
      required: true
      schema:
        type: integer
      description: Query ID
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not authenticated"
            error_code: "AUTH_REQUIRED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Not enough permissions"
            error_code: "INSUFFICIENT_PERMISSIONS"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            detail: "Resource not found"
            error_code: "NOT_FOUND"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      operationId: registerUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        201:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        409:
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      operationId: loginUser
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        200:
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      operationId: refreshToken
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        200:
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        401:
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/forgot-password:
    post:
      tags: [Authentication]
      summary: Request password reset
      operationId: forgotPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        200:
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /auth/reset-password:
    post:
      tags: [Authentication]
      summary: Reset password
      operationId: resetPassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        200:
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        400:
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change password
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        200:
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        401:
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # ============================================
  # USERS
  # ============================================
...
  /queries:
    get:
      tags: [Queries]
      summary: List query history
      operationId: listQueries
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: database_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, executing, completed, failed, cancelled]
        - name: start_date
          in: query
          schema:
            type: string
            format: date-time
        - name: end_date
          in: query
          schema:
            type: string
            format: date-time
      responses:
        200:
          description: Query history retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryListResponse'
        401:
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Queries]
      summary: Execute natural language query
      operationId: executeQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryRequest'
      responses:
        200:
          description: Query executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/Unauthorized'
  /queries/{query_id}:
    get:
      tags: [Queries]
      summary: Get query details
      operationId: getQuery
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      responses:
        200:
          description: Query details retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
        404:
          $ref: '#/components/responses/NotFound'
  /queries/sql:
    post:
      tags: [Queries]
      summary: Execute raw SQL
      operationId: executeSQL
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SQLExecutionRequest'
      responses:
        200:
          description: SQL executed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        400:
          $ref: '#/components/responses/ValidationError'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
  /queries/{query_id}/cancel:
    post:
      tags: [Queries]
      summary: Cancel running query
      operationId: cancelQuery
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      responses:
        200:
          description: Query cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
  /queries/{query_id}/follow-up:
    post:
      tags: [Queries]
      summary: Follow-up question
      operationId: queryFollowUp
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FollowUpRequest'
      responses:
        200:
          description: Follow-up executed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryResponse'
        401:
          $ref: '#/components/responses/Unauthorized'
  /queries/{query_id}/explain:
    get:
      tags: [Queries]
      summary: Get query explanation
      operationId: explainQuery
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      responses:
        200:
          description: Query explanation
          content:
            application/json:
              schema:
                type: object
                properties:
                  explanation:
                    type: string
                  generated_sql:
                    type: string
        401:
          $ref: '#/components/responses/Unauthorized'
  /queries/{query_id}/optimize:
    post:
      tags: [Queries]
      summary: Get optimization suggestions
      operationId: optimizeQuery
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      responses:
        200:
          description: Optimization suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        description:
                          type: string
                        impact:
                          type: string
                          enum: [low, medium, high]
        401:
          $ref: '#/components/responses/Unauthorized'
  /queries/{query_id}/visualize:
    post:
      tags: [Queries]
      summary: Auto-generate visualization
      operationId: visualizeQuery
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/QueryIdPath'
      responses:
        200:
          description: Visualization suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  visualizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/VisualizationSuggestion'
        401:
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # CONVERSATIONS (Phase 3)
  # ============================================
...